Cipher.java is fixed now!
Firstly, I created a char variable called "encryptedChar" which holds the encrypted version of each char of the inputString in "encrypt" method, and then I filled the empty for loop in line 19 inside the "encrypt" method, so this method will return outputString which will be formed by the concatenation of encryptedChar with outputString in each iteration. Moreover, I applied the same process for the "decrypt" method by copying and pasting the content of "encrypt" method. I just changed the name of the char variable from "encryptedChar" to "decryptedChar" and the boolean of "true" in replaceChar method's parameter into "false". Lastly, I created the missing return statement of " return CIPHER_ALPHABET.charAt(i); " in replaceChar method.